refactor dev

+++ database/migrations/2026_01_13_153556_add_gateway_to_deposits_table.php
@@ -8,16 +8,20 @@ return new class extends Migration {
     public function up(): void
     {
         Schema::table('deposits', function (Blueprint $table) {
-            $table->string('gateway')
-                  ->nullable()
-                  ->after('amount_kobo'); 
+            if (!Schema::hasColumn('deposits', 'gateway')) {
+                $table->string('gateway')
+                      ->nullable()
+                      ->after('amount_kobo');
+            }
         });
     }
 
     public function down(): void
     {
         Schema::table('deposits', function (Blueprint $table) {
-            $table->dropColumn('gateway');
+            if (Schema::hasColumn('deposits', 'gateway')) {
+                $table->dropColumn('gateway');
+            }
         });
     }
 };
